pyproject.toml
[project]
name = "rpiosc"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "PySide6==6.10.2",
  "pyqtgraph==0.14.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"
